Delete README.md
This is one of my favorite projects as, I personally figured out how to create all of the coding alone at first just using vanilla JS. i was very proud of my novice work and preserving through almost a month of fighting with this. I finally was able to reach out for help and finally getting this resolved without redoing my whole project. 

This is a perfect example of novice coder at work. The journey continues...